Changed atom species to be specified by simple string rather than an enumerated id which was causing an unnecessary proliferation of AtomSpeciesDatabase objects which were needed to convert from strings to ids.  This way is much cleaner.

+++ lib/spipe/tests/blocks/LowestFreeEnergyTest.cpp
@@ -92,7 +92,7 @@ BOOST_AUTO_TEST_CASE(LowestFreeEnergyTest)
   typedef Engine::RunnerPtr RunnerPtr;
 
   // SETTINGS /////////////
-  const unsigned int NUM_TO_KEEP    = 2;
+  const size_t NUM_TO_KEEP    = 2;
   const unsigned int NUM_STRUCTURES = 10;
 
   Pipe pipe;
